Estimation of the Weighted Cumulative Effect with Two-Level

This function estimates the effect of an exposure history on a health outcome (such as a binary event, survival time, or repeated event), using the WCIE (Weighted Cumulative Index of Exposure) approach. It is designed for longitudinal data where exposure varies over time, and where the goal is to model the effect of this time-varying exposure on an outcome.

A. Estimation follows the following steps:

(1) Individual prediction of exposure: Based on the mixed model mexpo, which must be an object of class hlme, individual exposure trajectories are predicted over a user-defined time window (timerange) and frequency (step). The presence or absence of a random intercept is taken into account, but random effects must be explicitly specified. The temporal structure (e.g., splines, polynomials) used in the original model is preserved, but must be directly included in the mexpo model.

(2) Reconstruction of the exposure history: A temporal weighting basis is used to construct the WCIE, i.e., time-weighted exposure scores. Currently, only natural splines (NS) are implemented. The spline knots are automatically computed based on time quantiles, according to the number of internal knots (knots) specified by the user.

(3) Computation of the WCIE: The WCIE corresponds to the weighted sum of the products between time spline basis functions and predicted exposure values. Each spline generates a separate WCIE component (e.g., WCIE1, WCIE2, ...).

(4) Outcome modeling: A model is then fitted to the outcome, incorporating the WCIE components as explanatory variables. Two types of models are currently available via the reg.type argument: - "logistic": logistic regression (GLM model) - "cox": Cox proportional hazards model (under development) The final formula is automatically reconstructed based on the original model, replacing the exposure term with the sum of WCIE components.

The function output includes: - the fitted outcome model (glm), - the temporal predictions (Ypred) and random effects, - the time quantiles used for splines, - the final formula used in the outcome model.

B. USAGE PRECAUTIONS:

(1) The estimation of WCIE relies on the quality of the exposure model provided in mexpo. Since this model is used to simulate exposure histories, incorrect specification of fixed or random effects may bias the WCIE estimates. It is strongly recommended to use natural splines (NS) for the time structure and ensure that returndata = TRUE is specified in the hlme object.

(2) The weighting function used to estimate the WCIE over time can lead to instability or overfitting if the number of knots is too high or if the time window (timerange) is poorly chosen. Users should inspect the shape of the weight function and consider simplifying it if convergence issues or unreasonable effects are detected.

(3) Convergence problems can occur in the outcome model if the WCIE variables are highly collinear, especially when using a fine time grid (step) or too many knots. If the model fails to converge, consider reducing the number of time points or the complexity of the spline basis.

Usage

WCIE2F(
  mexpo,
  var.time,
  times,
  weightbasis,
  knots,
  knots.vector,
  data,
  reg.type,
  model
)

Arguments

mexpo

An object of class hlme from the lcmm package, used to model the exposure process. The object must be created using the hlme function with specific arguments provided. The fixed effects formula and the random effects formula must be specified. The subject argument must indicate the subject ID, and the dataset must be provided via the data argument. It is essential to include returnData = TRUE in the function call to ensure that the internal data can be accessed.

var.time

character indicating the name of the time variable for also the exposition and the outcome data in the model mexpo.

times

Numeric vector of length 4 indicating the desired time window for exposure (min, max, step, alea).

weightbasis

Type of temporal weighting function used to estimate the Weighted Cumulative Indirect Effects (WCIE). This specifies the functional form used to model the influence of past exposures over time. Currently, the following options are available: "NS" for natural splines (implemented), "BS" for B-splines (to be developed), and "PS" for P-splines (to be developed).

knots

number of internal knots

knots.vector

list of 2 vector : one for the internal knots for the splines (used only for splines temporal weighting function) and a second for the boundary knots. (ex:knots.vector=list(knots=c(-15,-5),boundary.knots=c(-20,0)))

data

A data frame containing the variables specified in the outcome model model including the outcome variable. This dataset will be used to estimate the outcome model, and the WCIE variables calculated previously will be added to this dataset prior to model fitting.

reg.type

Type of outcome model: "logistic", "cox", ...

model

two-sided linear formula object for the outcome model. The response outcome is on the left of ~ and the covariates are separated by + on the right of ~. To include the effect of past exposure, you must explicitly add WCIE (or interaction terms such as WCIE:sex) to the formula. For example, Y ~ WCIE + age + sex or Y ~ WCIE:sex + age are valid formulas.

Value

A list containing:

model

outcome model object

data_expo

Intermediate data set with individual predictions.

mexpo

Exposure model hlme object provided at the beginning.

data_outcome

Data set used to fit the outcome model.

call

The matched call for the outcome model.

splines.quantiles

The internal quantiles used for the natural splines, which define the time-weighting function.

boundary.quantiles

The lower and upper bounds of the natural splines (5th and 95th percentiles), used to define the limits of the time-weighting function.

AIC

AIC criteria of the outcome model.

loglike

log-likelihood of the outcome model.

References

Maud Wagner et al. “Time-varying associations between an exposure history and a subsequent health outcome : a landmark approach to identify critical windows”. In : BMC Med Res Methodol (2021). doi : 10.1186/s12874-021-01403-w

See also

Author

Encore un giga beau gosse