Bootstrap Estimation of WCIE for a Single Replicate

This internal function is used to perform one iteration of the bootstrap procedure for estimating WCIE and fitting the corresponding outcome model using perturbed parameters.

Usage

doOneBootWCIE(
  boot_params,
  i,
  times,
  mexpo,
  var.time,
  weightbasis,
  knots,
  knots.vector,
  data,
  reg.type,
  model
)

Arguments

boot_params

A matrix or data frame containing all bootstrap parameter vectors.

i

Integer. The row index corresponding to a bootstrap sample in boot_params.

times

Numeric vector of length 4 indicating the desired time window for exposure (min, max, step, alea).

mexpo

An object of class hlme from the lcmm package, used to model the exposure process. The object must be created using the hlme function with specific arguments provided. The fixed effects formula and the random effects formula must be specified. The subject argument must indicate the subject ID, and the dataset must be provided via the data argument. It is essential to include returnData = TRUE in the function call to ensure that the internal data can be accessed.

var.time

character indicating the name of the time variable in the model mexpo.

weightbasis

Type of temporal weighting function used to estimate the Weighted Cumulative Indirect Effects (WCIE). This specifies the functional form used to model the influence of past exposures over time. Currently, the following options are available: "NS" for natural splines (implemented), "BS" for B-splines (to be developed), and "PS" for P-splines (to be developed).

knots

number of internal knots

knots.vector

list of 2 vector : one for the internal knots for the splines (used only for splines temporal weighting function) and a second for the boundary knots. (ex:knots.vector=list(knots=c(-15,-5),boundary.knots=c(-20,0)))

data

A data frame containing the variables specified in the outcome model model including the outcome variable. This dataset will be used to estimate the outcome model, and the WCIE variables calculated previously will be added to this dataset prior to model fitting.

reg.type

Type of outcome model: "logistic", "cox", ...

model

two-sided linear formula object for the outcome model. The response outcome is on the left of ~ and the covariates are separated by + on the right of ~. To include the effect of past exposure, you must explicitly add WCIE (or interaction terms such as WCIE:sex) to the formula. For example, Y ~ WCIE + age + sex or Y ~ WCIE:sex + age are valid formulas.

Value

A list containing:

coef

Estimated coefficients of the outcome model.

V

Variance-covariance matrix (intra + inter) of the estimators.

AIC

AIC criteria of the outcome model.

loglike

log-likelihood of the outcome model.

Details

The function takes a set of bootstrap parameters from boot_params, reconstructs the exposure model using those parameters, and then estimates the WCIE values. These WCIE components are then used to fit the outcome model. It returns the estimated regression coefficients, their variance-covariance matrix, the AIC, and the log-likelihood of the outcome model.