Data simulation according to different outcome models from WCIEpackage

This function simulates two sample, one exposition sample according to a model estimated with hlme, functions and an outcome sample simulate in relation with the effect of the exposition.

Usage

simulateWCE(
  object,
  nsim = 1,
  seed = NULL,
  times,
  internal.step,
  tname,
  n,
  Xbin = NULL,
  Xcont = NULL,
  weightbasis,
  reg.type,
  knots = NULL,
  knots.vector = NULL,
  model,
  coef.wcie,
  Xoutcome,
  Surv.param = NULL
)

Arguments

object

an object of class hlme

nsim

simulation number

seed

the random seed

times

either a data frame with 2 columns containing IDs and measurement times, or a vector of length 4 specifying the minimal and maximum measurement times, the spacing between 2 consecutive visits and the margin around this spacing

internal.step

Time increment used for high-resolution simulation before resampling to the user's desired time step.

tname

the name of the variable representing the measurement times in object. Default to the second column's name of times if it is a data frame, and to object$var.time otherwise.

n

number of subjects to simulate*

Xbin

an optional named list giving the probabilities of the binary covariates to simulate. The list's names should match the binary covariate's names used in object.

Xcont

an optional named list giving the mean and standard deviation of the Gaussian covariates to simulate. The list's names should match the continuous covariate's names used in object.

weightbasis

Type of temporal weighting function used to estimate the Weighted Cumulative Indirect Effects (WCIE). This specifies the functional form used to model the influence of past exposures over time. Currently, the following options are available: "NS" for natural splines (implemented), "BS" for B-splines (to be developed), and "PS" for P-splines (to be developed).

reg.type

Type of outcome model: "logistic", "cox", ...

knots

number of internal knots

knots.vector

Vector of internal knots for the splines (used only for splines temporal weighting function).

model

two-sided linear formula object for the outcome model. The response outcome is on the left of ~ and the covariates are separated by + on the right of ~. To include the effect of past exposure, you must explicitly add WCIE to the formula. For example, Y ~ WCIE + age + sex are valid formulas.

coef.wcie

named list giving the parameters of the WCIE variables to simulate.

Xoutcome

named list giving the parameters of the intercept and the covariables in the logistic model to simulate de outcome variable. The list's names should match the intercept and covariable names used in object

Surv.param

named list giving the parameter lambda to simulate the survival time variable and the upper limit of the uniform distribution used to simulate censoring times. (ex. list(lambda=1,max_censoring_time=2.5))

Value

exposition.data

a data frame with one line per observation and one column per variable. Variables appears in the following order : subject id, measurement time, entry time, binary covariates, continuous covariates, longitudinal outcomes.

outcome.data

a data frame with one line per individual and one column per variable. Variables appears in the following order : subject id, binary covariates, continuous covariates, binary outcomes.

Details

to do